Move InMemoryFileIO tests also.
